requirements.txt
streamlit
pandas
plotly
spacy
sentence-transformers
vaderSentiment
language-tool-python
python-docx
PyPDF2
